# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "EconomicResultRead": {
        "additionalProperties": false,
        "properties": {
          "amounts": {
            "$ref": "#/components/schemas/EconomicAmounts"
          },
          "assessment_status": {
            "const": "hold",
            "title": "Assessment Status",
            "type": "string"
          },
          "calculation_status": {
            "enum": [
              "conditional_user_assumption",
              "hold"
            ],
            "title": "Calculation Status",
            "type": "string"
          },
          "decision_at_utc": {
            "format": "date-time",
            "title": "Decision At Utc",
            "type": "string"
          },
          "economic_result_id": {
            "title": "Economic Result Id",
            "type": "string"
          },
          "evidence_level": {
            "const": "assumed",
            "title": "Evidence Level",
            "type": "string"
          },
          "formula_version": {
            "title": "Formula Version",
            "type": "string"
          },
          "hold_reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Hold Reason Codes",
            "type": "array"
          },
          "input_origin": {
            "const": "user",
            "title": "Input Origin",
            "type": "string"
          },
          "market_context_kind": {
            "const": "unavailable",
            "title": "Market Context Kind",
            "type": "string"
          },
          "market_hold_report_id": {
            "format": "uuid",
            "title": "Market Hold Report Id",
            "type": "string"
          },
          "market_scenario_result_id": {
            "title": "Market Scenario Result Id",
            "type": "string"
          },
          "quantities": {
            "$ref": "#/components/schemas/EconomicQuantities"
          },
          "sales_totals_status": {
            "enum": [
              "inventory_reconciled",
              "unverified_input_arithmetic"
            ],
            "title": "Sales Totals Status",
            "type": "string"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_revision": {
            "title": "Scenario Revision",
            "type": "string"
          }
        },
        "required": [
          "economic_result_id",
          "market_scenario_result_id",
          "scenario_id",
          "scenario_revision",
          "decision_at_utc",
          "formula_version",
          "market_context_kind",
          "market_hold_report_id",
          "calculation_status",
          "assessment_status",
          "sales_totals_status",
          "input_origin",
          "evidence_level",
          "quantities",
          "amounts",
          "hold_reason_codes"
        ],
        "title": "EconomicResultRead",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ServiceBearer": {
        "description": "Opaque server-issued service credential over HTTPS. Tenant and scopes are server-owned.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "OpenSmartFarmSim",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/break-even-results": {
      "get": {
        "operationId": "getBreakEvenResult",
        "parameters": [
          {
            "in": "query",
            "name": "plan_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "pattern": "^[^\\s\\x00-\\x1f](?:[^\\x00-\\x1f]*[^\\s\\x00-\\x1f])?$",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakEvenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Get Break Even Result",
        "x-ossf-required-scopes": [
          "break_even_read"
        ]
      }
    },
    "/v1/economic-results/{result_id}": {
      "get": {
        "operationId": "getEconomicResult",
        "parameters": [
          {
            "in": "path",
            "name": "result_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Result Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomicResultRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Get Economic Result",
        "x-ossf-required-scopes": [
          "market_result_read"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Get Job",
        "x-ossf-required-scopes": [
          "metadata"
        ]
      }
    },
    "/v1/jobs/{job_id}/hold-report": {
      "get": {
        "operationId": "getJobHold",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobHoldStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Get Job Hold Report",
        "x-ossf-required-scopes": [
          "metadata",
          "artifact",
          "auditor"
        ]
      }
    },
    "/v1/locations": {
      "post": {
        "operationId": "registerLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "goal_id": {
                    "maxLength": 200,
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:/-]{0,199}$",
                    "title": "Goal Id",
                    "type": "string"
                  },
                  "idempotency_key": {
                    "maxLength": 200,
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:/-]{0,199}$",
                    "title": "Idempotency Key",
                    "type": "string"
                  },
                  "latitude": {
                    "maximum": 90.0,
                    "minimum": -90.0,
                    "title": "Latitude",
                    "type": "number"
                  },
                  "longitude": {
                    "maximum": 180.0,
                    "minimum": -180.0,
                    "title": "Longitude",
                    "type": "number"
                  },
                  "period_end_utc": {
                    "format": "date-time",
                    "maxLength": 27,
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,6})?Z$",
                    "title": "Period End Utc",
                    "type": "string"
                  },
                  "period_start_utc": {
                    "format": "date-time",
                    "maxLength": 27,
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,6})?Z$",
                    "title": "Period Start Utc",
                    "type": "string"
                  }
                },
                "required": [
                  "latitude",
                  "longitude",
                  "period_start_utc",
                  "period_end_utc",
                  "goal_id",
                  "idempotency_key"
                ],
                "title": "LocationRequest",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Post Location",
        "x-ossf-max-body-bytes": 4096,
        "x-ossf-required-scopes": [
          "location_create"
        ]
      }
    },
    "/v1/market-hold-reports/{report_id}": {
      "get": {
        "operationId": "getMarketHold",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketHoldStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Get Market Hold Report",
        "x-ossf-required-scopes": [
          "market_hold_read"
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "getRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^synthetic-thermal-v1:[0-9a-f]{64}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalRunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "summary": "Get Run",
        "x-ossf-required-scopes": [
          "thermal_run_read"
        ]
      }
    },
...
  }
}